+++ lib/date_chips/date_chip.dart
@@ -20,20 +20,22 @@ class DateChip extends StatelessWidget {
   }) : super(key: key);
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(
-        top: 7,
-        bottom: 7,
-      ),
-      child: Container(
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.all(Radius.circular(6)),
-          color: color,
+    return Center(
+      child: Padding(
+        padding: const EdgeInsets.only(
+          top: 7,
+          bottom: 7,
         ),
-        child: Padding(
-          padding: const EdgeInsets.all(5.0),
-          child: Text(
-            Algo.dateChipText(date),
+        child: Container(
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.all(Radius.circular(6)),
+            color: color,
+          ),
+          child: Padding(
+            padding: const EdgeInsets.all(5.0),
+            child: Text(
+              Algo.dateChipText(date),
+            ),
           ),
         ),
       ),